pyproject.toml
[project]
name = "autosteer"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "torch>=2.0",
    "transformers>=4.40",
    "transformer-lens>=2.0",
    "anthropic>=0.40",
    "fire>=0.5",
    "numpy>=1.24",
    "pydantic>=2.0",
    "python-dotenv>=1.0",
    "scikit-learn>=1.3",
    "matplotlib>=3.8",
    "plotly>=5.18",
    "kaleido>=0.2",
    "modal>=0.73",
]
